% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{state_coords}
\alias{state_coords}
\title{state_coords}
\format{A data frame with 51 rows and 4 variables}
\description{
simple coordinates for states to use in a 'statebin' plot
\itemize{
  \item \code{abbrev}. state abbreviation, including \code{DC} (chr)
  \item \code{state}. capitalized state name (chr)
  \item \code{col}. colum where it fits (12x8 grid) (chr)
  \item \code{row}. row where it fits (12x8 grid) (chr)
}
}
\note{
Last updated 2014-08-25
}
\seealso{
\itemize{
  \item WaPo - \url{http://www.washingtonpost.com/wp-srv/special/business/states-most-threatened-by-trade/}
}
}
\keyword{datasets}

