% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statebins-package.R
\docType{package}
\name{statebins-package}
\alias{statebins-package}
\title{statebins is an alternative to choropleth maps for US States}
\description{
statebins is an alternative to choropleth maps for US States
}
\author{
Bob Rudis (@hrbrmstr)
}

