% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{statebins-package}
\alias{statebins-package}
\title{statebins is an alternative to choropleth maps for US States}
\description{
statebins is an alternative to choropleth maps for US States
}
\author{
Bob Rudis (@hrbrmstr)
}

