% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statebins-package.R
\docType{package}
\name{statebins-package}
\alias{statebins-package}
\title{Create United States Uniform Cartogram Heatmaps}
\description{
The cartogram heatmaps generated by the included methods
are an alternative to choropleth maps for the United States
and are based on work by the Washington Post graphics department in their report
on "The states most threatened by trade"
(\url{http://www.washingtonpost.com/wp-srv/special/business/states-most-threatened-by-trade/}).
"State bins" preserve as much of the geographic placement of the states as
possible but have the look and feel of a traditional heatmap. Functions are
provided that allow for use of a binned, discrete scale, a continuous scale
or manually specified colors depending on what is needed for the underlying data.
}
\author{
Bob Rudis (bob@rud.is)
}
