% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{package}
\name{statebins}
\alias{statebins}
\alias{statebins-package}
\title{A package to ...}
\usage{
statebins(state_data, state_col = "state", value_col = "value",
  text_color = "white", state_border_col = "white", breaks = 5,
  labels = 1:4, legend_title = "Legend", legend_position = "top",
  brewer_pal = "PuBu", plot_title = "", title_position = "bottom")
}
\description{
A package to ...

This version uses a discrete scale, binned by the "breaks" parameter
}
\author{
Bob Rudis (@hrbrmstr)
}

