% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{statebins_continuous}
\alias{statebins_continuous}
\title{Create ggplot-based "statebin" charts in the style of \url{http://www.washingtonpost.com/wp-srv/special/business/states-most-threatened-by-trade/}}
\usage{
statebins_continuous(state_data, state_col = "state", value_col = "value",
  text_color = "white", state_border_col = "white",
  legend_title = "Legend", legend_position = "top", brewer_pal = "PuBu",
  plot_title = "", title_position = "bottom")
}
\description{
This version uses a discrete scale, binned by the "breaks" parameter
}

