% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-statebin.R
\name{theme_statebins}
\alias{theme_statebins}
\title{Base statebins theme}
\usage{
theme_statebins(legend_position = "bottom", base_size = 11, base_family = "")
}
\arguments{
\item{legend_position}{fills in \code{legend.position}}

\item{base_family, base_size}{same as \code{theme_bw()}}
}
\description{
Clears out most of the cruft. Builds off of \code{theme_bw()}
}
